fix: compare result of `G_N_ELEMENTS()` with `size_t`
I've been told the actual result of `G_N_ELEMENTS()` is actually
`size_t`, rather than e.g. `unsigned int`, so fix all instances of
the incorrect comparison.

Additionally, don't use a static variable to store the length of
test arrays, since `G_N_ELEMENTS()` is a pre-processor macro anyways.

+++ src/libvalent/clipboard/valent-clipboard.c
@@ -484,7 +484,7 @@ valent_clipboard_read_text (ValentClipboard     *clipboard,
 
   if (mimetypes != NULL)
     {
-      for (unsigned int i = 0; i < G_N_ELEMENTS (text_mimetypes); i++)
+      for (size_t i = 0; i < G_N_ELEMENTS (text_mimetypes); i++)
         {
           const char *text_mimetype = text_mimetypes[i];
